+++ drivers/media/platform/nxp-vpu/nx_vpu_encoder.c
@@ -92,7 +92,7 @@ int NX_VpuEncOpen(struct vpu_open_arg *pOpenArg, void *devHandle,
 	hInst->devHandle = devHandle;
 
 	hInst->instBufPhyAddr = (uint64_t)pOpenArg->instanceBuf.phyAddr;
-	hInst->instBufVirAddr = (uint64_t)pOpenArg->instanceBuf.virAddr;
+	hInst->instBufVirAddr = (unsigned long)pOpenArg->instanceBuf.virAddr;
 	hInst->instBufSize    = pOpenArg->instanceBuf.size;
 
 	pEncInfo = &hInst->codecInfo.encInfo;
@@ -633,7 +633,8 @@ static int VPU_EncSetFrameBufCommand(struct nx_vpu_codec_inst *pInst,
 	}
 
 	swap_endian((unsigned char *)frameAddr, sizeof(frameAddr));
-	NX_DrvMemcpy((void *)pInst->paramVirAddr, frameAddr, sizeof(frameAddr));
+	NX_DrvMemcpy((void *)(unsigned long)pInst->paramVirAddr, frameAddr,
+			sizeof(frameAddr));
 
 	/* Tell the codec how much frame buffers were allocated. */
 	VpuWriteReg(CMD_SET_FRAME_BUF_NUM, pEncInfo->minFrameBuffers);
@@ -768,7 +769,7 @@ static int VPU_EncGetHeaderCommand(struct nx_vpu_codec_inst *pInst,
 	pEncInfo->strmWritePrt = wdPtr;
 	pEncInfo->strmReadPrt = rdPtr;
 
-	*ptr = (void *)pEncInfo->strmBufVirAddr;
+	*ptr = (void *)(unsigned long)pEncInfo->strmBufVirAddr;
 	*size = headerSize;
 
 	return VPU_RET_OK;
